OVManager security hardening, auth refresh, and UI/UX overhaul
Security:
- P0 fixes: CORS default same-origin, bot token Fernet encryption,
  traffic fields int/BigInteger, OVPN download ownership check,
  URLPATH injection fix, backup restore sanitization, Vite outDir,
  Python 3.12, Docker non-root, .dockerignore, .env.example hardened,
  SSL startup validation.
- P1: CSRF protection middleware, security headers middleware, pip-audit
  + Trivy in CI, pages-deploy uploads only frontend/dist.
- P2: UTC datetime fixes, concurrent daily checks, geolocation cache
  capped, refresh token flow (30m access / 7d refresh), axios
  auto-refresh interceptor.

UI/UX:
- Shared primitives: ErrorState, EmptyState, PanelSkeleton, StatusBadge
- ServerStats rewrite with error/empty/loading states
- UserManagement + NodeManagement: error states, empty states, badges
- LoginPage: loading state, double-submit guard
- OnboardingChecklist: progressive setup guide with live counts
- CommandPalette: Ctrl+K fuzzy search across pages/users/nodes
- Expiry chips, usage bars, QR codes in user detail modal
- OS theme sync on first visit, reduced-motion respected
- 21 new i18n keys

Tests: 10/10 passing. Build: clean.

+++ backend/schema/_input.py
@@ -5,8 +5,8 @@ from typing import Optional
 
 class CreateUser(BaseModel):
     name: str = Field(min_length=3, max_length=64)
-    total: Optional[float] = None
-    used: Optional[float] = None
+    total: Optional[int] = None
+    used: Optional[int] = None
     # Max simultaneous logins/devices per config. 1 = single login, 0 = unlimited.
     max_logins: int = Field(default=1, ge=0, le=1000)
     expiry_date: date
@@ -14,8 +14,8 @@ class CreateUser(BaseModel):
 
 class UpdateUser(BaseModel):
     name: str
-    total: Optional[float] = None
-    used: Optional[float] = None
+    total: Optional[int] = None
+    used: Optional[int] = None
     # Max simultaneous logins/devices per config. 1 = single login, 0 = unlimited.
     max_logins: Optional[int] = Field(default=None, ge=0, le=1000)
     expiry_date: Optional[date]
